+++ tiny-techies-frontend/src/AdminDashPages/AdminDashStudents.jsx
@@ -8,7 +8,6 @@ const AdminDashStudents = () => {
   const [studentsData, setStudentsData] = useState([]);
 
   useEffect(() => {
-    
     axios.get(`${process.env.REACT_APP_API_URL}/student/getAllStudentsAndCourses`)
       .then((response) => {
         setStudentsData(response.data.data);
@@ -19,16 +18,19 @@ const AdminDashStudents = () => {
   }, []);
 
   const handleDeleteStudent = (studentId) => {
-    
-    axios.delete(`${process.env.REACT_APP_API_URL}/user/delete/${studentId}`)
-      .then(() => {
-        setStudentsData(studentsData.filter(student => student.User_id !== studentId));
-        toast.success('Student Deleted successfully');
-      })
-      .catch((error) => {
-        console.error('Error deleting Student:', error);
-        toast.error('Failed to delete Student');
-      });
+    const confirmDelete = window.confirm('Are you sure you want to delete this student?');
+
+    if (confirmDelete) {
+      axios.delete(`${process.env.REACT_APP_API_URL}/user/delete/${studentId}`)
+        .then(() => {
+          setStudentsData(studentsData.filter(student => student.User_id !== studentId));
+          toast.success('Student Deleted successfully');
+        })
+        .catch((error) => {
+          console.error('Error deleting Student:', error);
+          toast.error('Failed to delete Student');
+        });
+    }
   };
 
   return (
@@ -37,40 +39,36 @@ const AdminDashStudents = () => {
         <div className="bgcardcourses">
           <div className="inputsdivcourses">
             <table border="1" className="tbcourse">
-              
-                <tr>
-                  <th>Name</th>
-                  <th>Age</th>
-                  <th>Email</th>
-                  <th>Course Name</th>
-                  <th>Absence</th>
-                  <th>Actions</th>
+              <tr>
+                <th>Name</th>
+                <th>Age</th>
+                <th>Email</th>
+                <th>Course Name</th>
+                <th>Absence</th>
+                <th>Actions</th>
+              </tr>
+              {studentsData.map((student) => (
+                <tr key={student.User_id}>
+                  <td>{student.UserFullName}</td>
+                  <td>{student.UserAge}</td>
+                  <td>{student.UserEmail}</td>
+                  <td>
+                    {student.EnrolledCourses && student.EnrolledCourses.length > 0 && (
+                      <ul>
+                        {student.EnrolledCourses.split(',').map((course, index) => (
+                          <li key={index}>{course}</li>
+                        ))}
+                      </ul>
+                    )}
+                  </td>
+                  <td>{student.UserAbsence}</td>
+                  <td>
+                    <button onClick={() => handleDeleteStudent(student.User_id)}>
+                      <img src="Images/trash-solid.svg" className="imgedit"  alt="Delete" />
+                    </button>
+                  </td>
                 </tr>
-             
-             
-                {studentsData.map((student) => (
-                  <tr key={student.User_id}>
-                    <td>{student.UserFullName}</td>
-                    <td>{student.UserAge}</td>
-                    <td>{student.UserEmail}</td>
-                    <td>
-                      {student.EnrolledCourses && student.EnrolledCourses.length > 0 && (
-                        <ul>
-                          {student.EnrolledCourses.split(',').map((course, index) => (
-                            <li key={index}>{course}</li>
-                          ))}
-                        </ul>
-                      )}
-                    </td>
-                    <td>{student.UserAbsence}</td>
-                    <td>
-                      <button onClick={() => handleDeleteStudent(student.User_id)}>
-                        <img src="Images/trash-solid.svg" className="imgedit"  alt="Delete" />
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-        
+              ))}
             </table>
           </div>
         </div>
